Serve CSS and woff2 for Poiret One.
Summary:
The WOFF 2.0 specification published by the Web Fonts Working Group
claims 30% average gain on WOFF 1.0, and all major browsers support
it, so this is the privileged format, with TTF as fallback.

The 2017 situation where a font must be served in EOT for IE,
SVG for old iOS, OTF, TTF and WOFF seems a story horror of the past.

To avoid flashing content, it's better to load a small piece of CSS
with the font information than hiding it in a Foundation huge CSS
blob compiled from SCSS. So we host CSS fonts definitions.

Ref T1881

Maniphest Tasks: T1881

Differential Revision: https://devcentral.nasqueron.org/D3135